Move tests back
