+++ packages/modern-di-fastapi/tests_fastapi/test_fastapi_di.py
