+++ packages/modern-di/tests_core/providers/test_context_adapter.py
