+++ packages/modern-di/tests_core/providers/test_resource.py
@@ -4,7 +4,7 @@ import typing
 import pytest
 from modern_di import Container, Scope, providers
 
-from tests.core.creators import create_async_resource, create_sync_resource
+from tests_core.creators import create_async_resource, create_sync_resource
 
 
 async_resource = providers.Resource(Scope.APP, create_async_resource)
