+++ packages/modern-di/tests_core/providers/test_selector.py
